Develop

+++ models/staging/gz_raw_data/stg_raw__ship.sql
@@ -1,21 +1,21 @@
-with 
+  with
 
-source as (
+  source as (
 
-    select * from {{ source('raw', 'ship') }}
+      select * from {{ source('raw', 'ship') }}
 
-),
+  ),
 
-renamed as (
+  renamed as (
 
-    select
-        orders_id,
-        shipping_fee,
-        logcost,
-        CAST(ship_cost AS float64) AS ship_cost
+      select
+          orders_id,
+          shipping_fee,
+          logcost,
+          CAST(ship_cost AS INT64) AS ship_cost
 
-    from source
+      from source
 
-)
+  )
 
-select * from renamed
+  select * from renamed
